app/pyproject.toml


[tool.poetry]
name = "FootballFantasyAPI"
version = "0.0.1"
description = "Fantasy Football API"
authors = ["YousefElmansy"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "~0.11.3"
fastapi = "~0.68.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "^3.0.1"
SQLAlchemy = "^1.4.22"
alembic = "^1.6.5"
tenacity = "^8.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
names = "^0.3.0"
pycountry = "^22.1.10"
randomname = "^0.1.5"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
pytest = "^7.0.0"
requests = "^2.27.1"
